requirements.txt
pyyaml>=6.0
pinecone>=5.0.0
